Handle null string in dump() JS function
The "null string" is a corner case of jsval, created by native code
using STRING_TO_JSVAL(NULL). This creates a jsval that has the type set
to string, but the string pointer set to NULL. Note that a null value
created by JS code (or JSVAL_NULL) has the type set to object.

Signed-off-by: Radu Rendec <radu.rendec@mindbit.ro>