Add js_log_error() function
This function replaces the error reporting mechanism from the old mozjs
implementation, namely the following functions/macros:
  * JS_ReportError
  * JS_ReportErrno
  * JS_RetError
  * JS_MiscErrorReporter
  * JS_MiscSetErrorReporter

In mozjs, all errors are reported using the engine error reporter, which
can be changed to a custom function using JS_SetErrorReporter. If native
C code calls a JS function that throws an error, the engine calls the
error reporter automatically and JS_CallFunctionName returns false.

In Duktape, things are slightly different. When the heap is created, a
custom fatal error handler can be assigned. An error thrown in JS code
invoked from a native C function can be fatal or not, depending on
whether the call is "protected", e.g. duk_safe_call() vs. duk_call(). If
the call is not "protected", the error is fatal and the engine calls the
fatal error handler automatically, which is supposed to terminate the
program. If the call is "protected", the C call returns DUK_EXEC_ERROR
and the error object is left at the top of the value stack. The purpose
of the js_log_error() function is to extract information from the error
object and log it.

Signed-off-by: Radu Rendec <radu.rendec@mindbit.ro>